move base exposure to uint32 (gives max 655.36 seconds, we will say 600s/10m

+++ exposuremode.go
@@ -11,7 +11,7 @@ type exposureMode struct {
 
 	paused        bool
 	running       bool
-	remainingTime int64
+	remainingTime int32
 }
 
 func newExpMode(s *stateData) *Mode {
@@ -33,7 +33,7 @@ func (e *exposureMode) SwitchTo(prev *Mode) {
 	// from the calling mode
 	e.prevMode = prev
 
-	e.remainingTime = int64(e.state.exposureSet.exposures[0].colTime[0]) * int64((10 * time.Millisecond))
+	e.remainingTime = int32(e.state.exposureSet.exposures[0].colTime[0]) * int32((10 * time.Millisecond))
 
 	e.running = true
 	e.state.SetLEDPanel(ledWhite)
@@ -45,7 +45,7 @@ func (e *exposureMode) SwitchAway() *Mode {
 	return e.prevMode
 }
 
-func (e *exposureMode) Tick(passed int64) (bool, bool) {
+func (e *exposureMode) Tick(passed int32) (bool, bool) {
 	if e.paused {
 		return false, false
 	}
@@ -92,7 +92,7 @@ func (e *exposureMode) UpdateDisplay(nextDisplay *[2][16]byte) {
 	nextDisplay[1] = stringTable[2][1]
 
 	if e.running {
-		num.Out(&nb, num.Num(e.remainingTime/int64((10*time.Millisecond))))
+		num.Out(&nb, num.Num(e.remainingTime/int32((10*time.Millisecond))))
 		copy(nextDisplay[1][12:16], nb[0:4])
 	}
 }
